Fix: renamed closet to UserCloset
